Refactor file structure and imports
